enowars3/explotify/src/requirements.txt
alembic==1.0.10
backcall==0.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cssselect==1.0.3
decorator==4.4.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
lxml==4.3.3
Mako==1.0.9
markovify==0.7.1
MarkupSafe==1.1.1
parsel==1.5.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py-espeak-ng==0.1.8
pydub==0.23.1
Pygments==2.4.0
pymongo==3.8.0
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.3
w3lib==1.20.0
wcwidth==0.1.7
Werkzeug==0.15.2
cachetools==2.1.0


